Move jetty server to its own package
