+++ java/server/src/org/openqa/selenium/grid/distributor/httpd/BUILD.bazel
@@ -16,6 +16,7 @@ java_library(
         "//java/server/src/org/openqa/selenium/grid/sessionmap",
         "//java/server/src/org/openqa/selenium/grid/sessionmap/config",
         "//java/server/src/org/openqa/selenium/grid/web",
+        "//java/server/src/org/openqa/selenium/jetty/server",
         "//third_party/java/beust:jcommander",
         "//third_party/java/guava",
         "//third_party/java/service",
