+++ java/server/src/org/openqa/selenium/grid/node/httpd/BUILD.bazel
@@ -18,6 +18,7 @@ java_library(
         "//java/server/src/org/openqa/selenium/grid/node/local",
         "//java/server/src/org/openqa/selenium/grid/server",
         "//java/server/src/org/openqa/selenium/grid/web",
+        "//java/server/src/org/openqa/selenium/jetty/server",
         "//third_party/java/beust:jcommander",
         "//third_party/java/guava",
         "//third_party/java/service",
