+++ java/server/src/org/openqa/selenium/grid/router/httpd/BUILD.bazel
@@ -17,6 +17,7 @@ java_library(
         "//java/server/src/org/openqa/selenium/grid/sessionmap",
         "//java/server/src/org/openqa/selenium/grid/sessionmap/config",
         "//java/server/src/org/openqa/selenium/grid/web",
+        "//java/server/src/org/openqa/selenium/jetty/server",
         "//third_party/java/beust:jcommander",
         "//third_party/java/service",
     ],
