+++ java/server/src/org/openqa/selenium/grid/server/BUILD.bazel
@@ -3,6 +3,7 @@ java_library(
     srcs = glob(["*.java"]),
     visibility = [
         "//java/server/src/org/openqa/selenium/grid:__subpackages__",
+        "//java/server/src/org/openqa/selenium/jetty/server:__pkg__",
         "//java/server/src/org/openqa/selenium/netty/server:__pkg__",
         "//java/server/src/org/openqa/selenium/remote/server:__pkg__",
         "//java/server/test/org/openqa/selenium/grid:__subpackages__",
