+++ java/server/src/org/openqa/selenium/grid/sessionmap/httpd/BUILD.bazel
@@ -14,6 +14,7 @@ java_library(
         "//java/server/src/org/openqa/selenium/grid/sessionmap",
         "//java/server/src/org/openqa/selenium/grid/sessionmap/local",
         "//java/server/src/org/openqa/selenium/grid/web",
+        "//java/server/src/org/openqa/selenium/jetty/server",
         "//third_party/java/beust:jcommander",
         "//third_party/java/guava",
         "//third_party/java/service",
