+++ java/server/src/org/openqa/selenium/grid/web/BUILD.bazel
@@ -5,6 +5,7 @@ java_library(
         "//java/server/src/org/openqa/selenium/grid:__subpackages__",
         "//java/server/src/org/openqa/selenium/remote/server:__subpackages__",
         "//java/server/test/org/openqa/selenium/grid:__subpackages__",
+        "//java/server/test/org/openqa/selenium/jetty/server:__pkg__",
     ],
     deps = [
         "//java/client/src/org/openqa/selenium:core",
