+++ java/server/src/org/openqa/selenium/jetty/server/BUILD.bazel
@@ -1,0 +1,17 @@
+java_library(
+    name = "server",
+    srcs = glob(["*.java"]),
+    deps = [
+        "//java/client/src/org/openqa/selenium:core",
+        "//java/client/src/org/openqa/selenium/json",
+        "//java/client/src/org/openqa/selenium/remote/http",
+        "//java/server/src/org/openqa/selenium/grid/server",
+        "//third_party/java/jetty",
+        "//third_party/java/servlet:javax.servlet-api",
+    ],
+    visibility = [
+        "//java/server/src/org/openqa/selenium/grid:__subpackages__",
+        "//java/server/test/org/openqa/selenium/jetty/server:__pkg__",
+        "//java/server/test/org/openqa/selenium/grid:__subpackages__",
+    ]
+)
