+++ java/server/test/org/openqa/selenium/jetty/server/BUILD.bazel
@@ -1,19 +1,16 @@
 load("//java:defs.bzl", "java_test_suite")
 
 java_test_suite(
-    name = "MediumTests",
-    size = "medium",
-    srcs = glob(["*Test.java"]),
+    name = "SmallTests",
+    size = "small",
+    srcs = glob(["*.java"]),
     deps = [
         "//java/client/src/org/openqa/selenium/json",
-        "//java/client/src/org/openqa/selenium/remote",
-        "//java/server/src/org/openqa/selenium/grid/config",
-        "//java/server/src/org/openqa/selenium/grid/server",
+        "//java/client/src/org/openqa/selenium/remote/http",
         "//java/server/src/org/openqa/selenium/grid/web",
         "//java/server/src/org/openqa/selenium/jetty/server",
         "//java/server/test/org/openqa/testing",
         "//third_party/java/assertj",
-        "//third_party/java/guava",
         "//third_party/java/junit",
         "//third_party/java/servlet:javax.servlet-api",
     ],
