+++ java/server/test/org/openqa/selenium/jetty/server/HttpHandlerServletTest.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.openqa.selenium.grid.server;
+package org.openqa.selenium.jetty.server;
 
 import org.junit.Test;
 import org.openqa.selenium.grid.web.ErrorCodec;
